Fixed the address manager for the new version of the wire package. Fixed errors in tests in wire and fixed a addrmgr.DeserializeNetAddress.
bmd.go now works with the updated versions of wire and addrmgr. Stream numbers are now uint32 everywhere for consistency.

+++ wire/message_test.go
@@ -59,8 +59,8 @@ func TestMessage(t *testing.T) {
 		t.Errorf("NewNetAddress: %v", err)
 	}
 	me_expected.Timestamp = time.Time{} // Version message has zero value timestamp.
-	msgVersion := wire.NewMsgVersion(me, you, 123123, []uint64{1})
-	msgVersionExpected := wire.NewMsgVersion(me_expected, you_expected, 123123, []uint64{1})
+	msgVersion := wire.NewMsgVersion(me, you, 123123, []uint32{1})
+	msgVersionExpected := wire.NewMsgVersion(me_expected, you_expected, 123123, []uint32{1})
 
 	msgVerack := wire.NewMsgVerAck()
 	msgAddr := wire.NewMsgAddr()
